--- src/OverlayTrigger/index.js
@@ -1,6 +1,0 @@
-import OverlayTrigger, { OVERLAY_TRIGGER_PLACEMENT } from './OverlayTrigger';
-
-export {
-  OverlayTrigger,
-  OVERLAY_TRIGGER_PLACEMENT,
-};
