+++ src/OverlayTrigger/index.ts
@@ -1,0 +1,6 @@
+import { OverlayTrigger, OVERLAY_TRIGGER_PLACEMENT } from './OverlayTrigger';
+
+export {
+  OverlayTrigger,
+  OVERLAY_TRIGGER_PLACEMENT,
+};
